requirements.txt
# Core dependencies
psycopg2-binary>=2.9.0
imaplib2>=3.6
python-dotenv>=1.0.0
pydantic>=2.5.0
cryptography>=41.0.0

# Data processing
pytz>=2023.3

# Email handling
email-validator>=2.1.0

# Logging and monitoring
structlog>=23.2.0

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.12.1
flake8>=7.0.0
mypy>=1.8.0

# Security
bcrypt>=4.1.2
